Documentation update for Sample Backend (#56)
Documentation fixes for sample_backend

Updating documentation to align with the code and new screenshots for interface updates.

More updates to Postman section

Correcting the documentation to work with as-built.

Updates to last sections of instructions

Fixed inconsistencies in the documentation with the as-built structure of the Events and Endpoints resources in Postman. Additionally, the images and naming of the AWS IoT device was corrected.

Updated Postman Collection including minor changes and bump to 2.1 version.